Updated Registry files:
src/core_atmosphere/Registry.xml
src/core_init_atmosphere/Registry.xml

to include multigrid varibles and config options. Both Registry files have been edited in the same way and amendments are detailed below. Some of the chnages arent required for the init code as the multi-grid is not implemented but since it passes through mostly the same code, these need to be defined unless further changes are made to work around this. Summary of the changes made:

>> For both init and model run cores:

[a] Added config_num_mgslevels to "nhyd_model" as an integer {default = 1}
[b] Added config_mg_input_name below in "io" as a character {default = "mg_grid.nc"}

>> For just the model run core
[c] Added EIGHT to dims
[d] Added config_alpha_si {default = 0.5} and config_mg_solve_npass {default = 1} to "nhyd_model"
[e] Added variables to diag: pressure_inc and rhs
[f] Added variables to tend: tend_theta_fast, tend_rho_fast, tend_w_fast and tend_u_fast
[g] Added a new data structure called 'mgs_vars' based on 'diag_physics' with these varibles: helm_inc, helm_rhs, restrictStencil, restrictWeights and helm_coeff (uses dim EIGHT)