Add new module for producing soundings from MPAS-Atmosphere
The sounding locations to be written are specified in a text file named
'sounding_locations.txt', which should contain three columns giving, respectively,
the latitudes (in degrees), the longitudes (in degrees), and the names (no whitespace)
of the sounding locations; if this file is not present, no soundings are written.

The interval between soundings is specified in the namelist record &soundings
using the config_sounding_interval variable.

Soundings are written to files named <sounding_location_name>.<timestamp>.snd,
where <sounding_location_name> is the name specified in the third column of the
sounding_locations.txt file, and <timestamp> is of the form YYYYMMDDHHmmss.