Add a trailing '/' as needed to config_geog_data_path in GWDO field computation
The code for the static field (terrain, land use, soil category, etc.) computation
adds a trailing '/' character to the config_geog_data_path specified in the namelist
if one isn't present. This commit adds the same logic in the code for computing
the GWDO static fields.

Thanks to J. Cipriani (TWC/IBM) for proposing these changes.