Example of adding a namelist-controlled number of passive tracers
The changes here permit a variable number (currently, up to 10) of passive
tracers to be added to a simulation by increasing the value of a new
namelist option, config_num_passive_tracers, to values between 1 and 10.

The new passive tracers are named tracer{1..10}, and each is controlled
by its own package, so that passive tracers beyond those requested in
the namelist are not allocated and do not appear in output files.

The tracers are initialized to a sinusoidally varying pattern with values
between 0 and 1, but other initial fields can be provided in the new
atm_init_tracers routine in mpas_atm_core.F.

Being passive, the additional tracers do not influence the model results,
absent any code optimizations by the compiler.