Fix issue in deallocating fields with multiple time levels and multiple blocks
When the keep_head_ptr argument to mpas_deallocate_field was .true.,
the mpas_deallocate_field routines previously attempted to deallocate all blocks
for the input field besides the first block. This led to problems in freeing
fields with multiple blocks and multiple time levels, since the blocks of
the field were each elements of an array of fields (the array containing
the time levels).

Now, setting the optional argument keep_head_ptr to .true. in a call to
one of the mpas_deallocate_field routines will prevent the containing
field type for all blocks from being deallocated. Memory leaks are still
avoided, since the field arrays for all blocks should eventually be
deallocated (e.g., with calls to mpas_deallocate_block in the loop
over blocks in lines 124-128 of mpas_domain_routines.F).