Add fake code to test ezxml library and update CMakeLists
The main program now calls C code in the src/framework directory to
parse a dummy string with the ezxml library.

In order to add the ezxml library and include paths to the build,
the global properties MPAS_INTERNAL_INCLUDE_DIRS and MPAS_INTERNAL_LIBRARIES
are set in the src/external/ezxml/CMakeLists.txt file and obtained
in the top-level CMakeLists.txt file, where they are used as arguments
to include_directories and target_link_libraries.

The ezxml library itself is added to the project with an add_library
command in src/external/ezxml/CMakeLists.txt file.