Add simplified atmosphere core that depends on Registry-generated code
The main program now depends on a simple atmosphere core module, which in turn
includes source code that is auto-generated by the Registry mechanism.

The Registry code generator is added as a new executable to the project, and
it parses a temporary Registry_temp.xml file in the core_atmosphere directory.

Eventually, the Registry_temp.xml file should be generated by pre-processing
the Registry.xml file.