Add CMakeLists.txt files for complete build of the atmosphere core.
In this commit, the name of the main program has been changed from "mpas"
to "mpas_main", since the MPAS physics require the macro "mpas" to be defined,
and the compilation macros needed by the atmosphere core are added as target
compilation defintions that are used across the entire project. Ideally,
we could set compilation definitions just for the source files in a core.