Define MPAS_NAMELIST_SUFFIX, MPAS_GIT_VERSION, and MPAS_EXE_NAME for parse
The Registry code generator makes use of several preprocessing macros to
generate Fortran code used in the core interface module. These macros are
now defined in the CMakeLists.txt for the parse program.

Thanks to https://bravenewmethod.com/2017/07/14/git-revision-as-compiler-definition-in-build-with-cmake/
for describing how to capture the output of a git command into a CMake
variable.