Add example time loop to dummy_core_run(...)
This commit adds a new namelist option, config_time_step, to enable the
specification of the time step for the model clock; previously, the time step
was always set to be equal to the difference between the start and stop times.

Additionally, a simple example of a time loop from the clock starting time to
the clock stopping time has been added to the dummy_core_run routine.