Examples of how to instantiate fields at runtime with stream I/O support
In order for fields that are instantiated at runtime to be read/written
in runtime-defined streams, the fields must be instantiated before the stream
manager is initialized (or before the streams XML file is parsed?).
Further, these fields must be instantiated after (SOMETHING).

A convenient place to instantiate new fields not defined in the Registry.xml
file is therefore at the end of each core's setup_block routine.

This commit provides two examples:
1) A simple, 2-d, time-invariant field; and
2) A var_array with three constituents

Not included in the var_array example are dimensions to identify the indices
of each constituent, or the dimensions that identify the start and end of
each array_group.