Use FIFO "ticket lock" in SMIOL async code to improve performance
With the Intel compilers, the main SMIOL I/O thread was apparently having
problems acquiring the mutex in order to add items to the write queue.

This commit adds code for a FIFO "ticket lock" (stolen from stackoverflow;
referenced in the code comments) to improve fairness of access to the queue.

Also included in this commit are a few cleanup items: removing unnecessary
scheduling priority and updating Fortran derived types to match C structs.