Don't rewind the namelist file before reading each namelist group
When MPAS is built with the NAG compiler, reads of namelist groups after
the first return a non-zero error code, which then causes MPAS to raise
a critical error. For example, with the 'test' core, the following critical
message is generated:

   Reading namelist from file namelist.test
  CRITICAL ERROR: Error while reading namelist record io.

This commit modifies the gen_inc.c code so that in the generated
namelist_defines.inc Fortran code the REWIND statement is commented-out, with
a comment above it explaining that it can be uncommented to allow namelist
groups to appear in any order. For example, the code to read the 'io' group
in the 'test' core now looks like the following:

        if (dminfo % my_proc_id == IO_NODE) then
  ! Rewinding before each read leads to errors when the code is built with
  ! the NAG Fortran compiler.
  ! If using other compilers, the rewind statement may be uncommented
  ! to allow namelist groups to appear in any order in a namelist file.
  !        rewind(unitNumber)
           read(unitNumber, io, iostat=ierr)
        end if
        call mpas_dmpar_bcast_int(dminfo, ierr)