Use an explicit format when converting dt to a string in mpas_set_timeInterval
If the real-valued 'dt' optional argument is given to the mpas_set_timeInterval
routine, the value of this variable is written into a string using an internal
write statement. Before this commit, list-directed output was used in this write
(i.e., the format was *), which permits the compiler to choose either the 'f'
or 'e' edit descriptors. Apparently, some compilers that choose an 'e' edit
descriptors encounter subsequent errors when trying to work with the time string.

This commit dynamically builds a format string that uses the 'f' edit descriptor.