Make the 'cam_topo' stream optional, and assume zero terrain if it isn't present
Previously, the 'cam_topo' stream was required for init case 13. However, there
are cases where a zero terrain field is needed. Rather than forcing users to
define a netCDF file with a zero PHIS field, this commit makes the 'cam_topo'
stream options. If this stream isn't present in the streams.init_atmosphere
file, the terrain field is set to zero.

This commit also improves the error message when config_specified_zeta_levels
isn't specified in the &vertical_grid namelist group, and it also adds print-out
of the min and max values of the terrain field used in the generation of
the vertical grid.