When MPAS-A is a CAM dycore, define tend_physics with tend_{ru,rtheta,rho}_physics
When MPAS-A is a CAM dycore, the tend_physics pool (var_struct) must be defined
to include tend_ru_physics, tend_rtheta_physics, and tend_rho_physics. This
commit modifies the Registry.xml file so that the tend_physics pool is always
defined, but its contents depend on whether DO_PHYSICS or MPAS_CAM_DYCORE are
defined when pre-processing the file.